fix db query
