+++ WebService/src/main/java/com/acmseproject/WebService/User/UserController.java
@@ -210,23 +210,4 @@ public class UserController {
             userRepository.changePassword(id, password);
         }
     }
-
-    /**
-     * POST-Method to change a user avatar
-     *
-     * @param key   API-Key for authentication
-     * @param id    User ID connected to the change
-     * @param image Image ID
-     */
-    @PostMapping(path = "/changeAvatar")
-    public void changeAvatar(@RequestParam String key, @RequestParam int id, @RequestParam int image) {
-        System.out.format("[Request] changeAvatar\n[Key] %s\n", key);
-        if (Objects.equals(key, userRepository.checkAuth(key))) {
-            System.out.format("[Verification] Valid\n");
-            String jsonString = new JSONObject()
-                    .put("image", Integer.toString(image))
-                    .toString();
-            userRepository.changeAvatar(id, jsonString);
-        }
-    }
 }
